GRID, IO and surfaceTemperature updated
The module surfaceTemperature.py has now only one function to calculate
the energy-balance fluxes, and one function for optimization.

The remove_melt_weq function in the GRID (grid.py) class has been
updated. Snowheight changes are now calculated by the water equivalent
and not the energy.

The tile option in the IO class (cpkernel/io.py) did only work for the
WRF grid and not for the regular lat/lon grid. This has been corrected.